pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "daita-cli"
version = "0.1.0"
description = "Daita CLI — manage and observe your hosted AI agents"
readme = "README.md"
authors = [{name = "Daita", email = "support@daita-tech.io"}]
license = {text = "Apache-2.0"}
requires-python = ">=3.11"
keywords = ["ai", "agents", "cli", "mcp", "daita"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "click>=8.0",
    "httpx>=0.27.0",
    "mcp>=1.0.0",
    "pyyaml>=6.0",
]

[project.scripts]
daita = "daita_cli.main:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.26",
    "respx>=0.21",
    "pytest-cov>=4.1",
]

[tool.setuptools.packages.find]
include = ["daita_cli*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["."]
addopts = "--tb=short -v"
